fix: Set redux-saga-thunk name correctly
In redux-saga-thunk v0.7.0 release a change was made to update the meta.thunk
value to be a object that holds name, key, type props with optional id prop.
Previously the action was setting the thunk name by meta.thunk = 'stringName'
and that was resulting in the meta.thunk object having properties for each char
in the thunk name string. This updates the actions to set the name by passing an
object like: meta.thunk = {name: 'stringName'}

NOTE: In the future we should update the name to be passed without the resource
path and instead set the name and id. Here is an example:

   ```js
   `meta.thunk = {
     name: 'DetailRead',
     id: 'user/123',
   }
   ```

This would allow for the new redux-saga-thunk selectors id arg to be used like
so:

   ```js
   pending(state, 'DetailRead', 'user/123')
   ```