feat: Add requestResult values to ResourceLoader child render prop obj
This adds  `entities` and `requestResult` props to the object passed to the child render prop of ResourceLoader component. This provides access to the raw request response action payload in `requestResult`. The `entities` prop will hold an entityId or entityIdList depending on if a list or detail was loaded. You can use `entities` with an EntityDetail or EntityList component to render results from the global state which is required if updates based on external actions is desired.