feat: ResourceLoader new resource request on prop changes & test helpers added (#38)
* feat: Update ResourceLoader prop name `autoLoad`

Because the ResourceLoader component will be updated in this PR to
request resource automatically when the resource prop changes it didn't
make sense to keep the name of `loadOnMount` prop. Renamed `loadOnMount`
prop to `autoLoad`.

* chore: Config Jest to require *.test.js file ext

I keep running into situations where I want to make special helper or
util configs for tests but when I add a file for that Jest failes
because there aren't test assertions. I rather just require .test.js
than put those test helper/util files outside of the `__tests__` folder.

* chore: Bum react-testing-library 2.3.0 to 3.1.3

* tests: Add standard test utils w/ renderWithRedux

Moved the slopy code at the top of the ResourceLoader.test.js to
src/utils/test/index.js so it can be used by all tests. Also added
custom return `rerender` function from `renderWithRedux` so it correctly
provides the redux store prop.

The `rerender` function is useful to test components behavior when their
props change. If you don't use `rerender` a new instance will be mounted
to new container.

Also create API in utils/test so we don't have to do that in all our
tests. `api` and `mockApi` are exported so the test modules can access
them.

* feat: ResourceLoader autoLoad on prop change

The resource helper component ResourceLoader now will respond to changes
to `resource` and `resourceId` props and load the resource if `autoLoad`
boolean prop is true. This should make route change based loading
automatic.

Also note that the old `loadOnMount` prop is replaced by `autoLoad`. If
you need to load the resource on mount you should use the `autoLoad`
bool prop flag.

Issue #37

* chore: Clean some test code for ResourceLoader

* Add @dyyylan as a contributor

* chore: Minor cleanup & added Dylan as contributor

* fix: Disable react lifecycle depreciation warning

* fix: Polyfill new React 16 lifecycles

Added `react-lifecycles-compat` to project to polyfill the
ResourceLoader component with the new `static getDerivedStateFromProps`
lifecycle. This allows us to support React 15 while coding with the
current best practices.

* chore: Fix prettier formatting to fix builds

There was an update to prettier in recent releases that results in
functional composition to format differently. I re-installed
`uptrend-scripts` locally and restarted `eslint_d` to get this
formatting correctly locally in my editor which means when I commit the
CI server will pass the lint step!

* test: Move test helpers to __tests__/helpers/

Previously I updated the jest config that `uptrend-scripts` provides to
a custom match pattern to only match files with .test. in the name. This
was to allow helper modules to be created next in the __tests__ folders
to be used by the actual test files. Setting that custom config worked
but when I tried to merge my PR it failed because the code coverage was
way to low due to the fact jest considered the helper file not tested.

Looking at the `uptrend-scripts` default jest config I realized that
`__tests__/helpers/` was defined to be ignored completely when looking
for tests and what files to create covereage reports for.
It does this by setting them in jest config option
`testPathIgnorePatterns` and `coveragePathIgnorePatterns`.

So the real solution was to add a `helpers/` folder in the `__tests__`
dir I created a helper in.

* fix: EntityDetail pass null so react can render

When there isn't a matching entity detail for the entityType and
entityId props passed in the component was passing `undefined` to the
child render prop function. This change returns `null` instead so you
can react can render it allowing the common render conditional:

  `entity && <ShowEntity entity={entity} />`

* test: Add tests for entites helpers