feat: ResourceLoader `updateResource` callback
ResourceLoader now passes `updateResource` function in the bag it passes
to child render prop. This new function triggers a resourceUpdateRequest
(PUT) for the resource it is setup to load. The `updateResource` takes a
single argument `data` which is the payload sent in the request. This
new function should make it easy to load a resource on page load and
then send updates (saves) to the backend with the response updating the
data rendered in the success view or child render prop.

Example Usage:

  <ResourceDetailLoader
    resource="user"
    resourceId={1}
    renderSuccess={user => <p>User: {result.name}</p>}
    autoLoad
  >
   {({
    status,
    result: user,
    updateResource,
   }) => (
     <div>
       {statusView}
       {status.success && (
         <button onClick={() => updateResource({...user, name: 'Ben'})}>
           Update User
         </button>
       )}
      </div>
    )}
  </ResourceDetailLoader>